Added Footsteps and attenuation
Footstep event is called footstep. There is also a switch event which should be used to set the footstep sample based on the material the player is walking on. There is also a state method called stealth which has two states, walk and sneak that change the volume of the footsteps.