Enemy Spawning and a shit ton of fixes
Added an enemy manager. It tracks the amount of enemies in a scene. If they fall below a threshold, it spawns more (until there are no more to spawn)
Added ability for enemies to die. They also broadcast their death the enemy manager.
Made it so that other enemies are alerted if one enemy detects the player or if the player shoots at enemy
Added a line renderer to show where the bullet it going

Fixed even more issues with the power source
Fixed issues with enemy health script